feat: Refactor Api planeacion_evaluacion_mid / https://github.com/udistrital/planeacion_cliente/issues/974
